requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0

# Dataset Download
kaggle>=1.5.0
kagglehub>=0.2.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Data Processing
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
